requirements.txt
Django==4.1.5
djangorestframework==3.12.1
psycopg2-binary==2.9.5
python-decouple==3.7

# Django	4.1.5	4.1.5
# asgiref	3.6.0	3.6.0
# djangorestframework	3.12.1	3.14.0
# pip	22.3.1	22.3.1
# pytz	2022.7.1	2022.7.1
# setuptools	65.5.1	67.0.0
# sqlparse	0.4.3	0.4.3
# wheel	0.38.4	0.38.4
